requirements.txt
xmpppy
pydns
-e hg+http://bitbucket.org/runeh/tweetstream/#egg=tweetstream
